Fix for #4, update README.md, clean up document_tags.py
In order to resolve #4 I have used starmap within the main iretator.
This also means that the main Batch job worker won't work anymore -
however, I am planning on removing the duplicate code in that function
and instead providing it as a CLI command from Flask. Will be opening
a new issue for that.